Updated solarized colors
I changed the color used for comments in solarized. This was done to
increase the contrast when I use gvim and solarized-light to look at code.
The old comment color was too light, making it hard to read the comments.
The new color is darker, although it is not as differentiated from the
color used for actual code. I am okay with this trade in performance. I
would rather be able to clearly read the comments. I can alwyas tell which
text is a comment in the code based on the language-appropriate syntax
for comments.

	modified:   vim/colors/solarized.vim