Added an alias for Jupyter Notebook
I created an alias for Jupyter Notebook: 'jn'

I also made 'gvim' act sanely in Linux. Most of the time, I log in
remotely to a linux machine, so I have 'gvim' pointed to 'vi'. On a Mac,
'gvim' is pointed to 'mvim', which is MacVim.

I also cleaned up the formatting of the file.

	modified:   bash_aliases