Commenting out the automatic SSH start on Linux machines
In bash_config, I have commented the lines that led to an auto start
for ssh-add. This only led to an annoying password prompt at every shell
login. I need to find a better way to handle this on a Linus machine.
For now, I will have to tolerate an SSH password prompt when I attempt
to use Git and GitHub together.

	modified:   bash_config