Added path information for the USGS ISIS software
I added path information and source a startup script for the USGS ISIS
software. ISIS is only being used on my deskstop right now, so I set
things up, using the 'hostname' command, so that the ISIS paths and
script are only used when on my office desktop machine, guabancex.

	modified:   bash_paths