Updated my bash profile to deal with an annoyance from conda
For some reason, the newest version of conda was modifying my prompt
when in the base, default environment. I keep my prompt compact for a
reason -- these things get out of control otherwise. I find it annoying
that conda is modifying my default environment prompt. Fortunately,
I found a simple fix for this, thanks to the issues section of the
conda GitHub repository.

	modified:   bash_profile