Hodgepodge of changes
Some more of the tweaks to the dotfiles that seem to be a constant task.
I updated paths and profile to reflect that my laptop, huracan, uses a
miniconda installation instead of an Anaconda installation. I also
updated the git difftool call in the global gitconfig. Finally, there
are a few other minor edits -- use git diff to see the changes.

 On branch master
 Your branch is up to date with 'origin/master'.

 Changes to be committed:
	modified:   bash_paths
	modified:   bash_profile
	modified:   gitconfig